Feat: #58- Delete API code, Add ViewModel
View역할의 SignInVC에서 카카오, 애플 로그인 API 코드를 제거하여 viewModel로 옮겨주었습니다.
